make utils tests unit tests
